% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aactr-package.R
\docType{package}
\name{aactr-package}
\alias{aactr}
\alias{aactr-package}
\title{aactr: Client for the ClinicalTrials.gov AACT API}
\description{
Client for the ClinicalTrials.gov AACT API: https://aact.ctti-clinicaltrials.org/.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maia-sh/aactr}
  \item Report bugs at \url{https://github.com/maia-sh/aactr/issues}
}

}
\author{
\strong{Maintainer}: Maia Salholz-Hillel \email{jane@example.com} (\href{https://orcid.org/0000-0003-1934-9504}{ORCID}) [copyright holder]

}
\keyword{internal}
