% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{aggregate_interventions}
\alias{aggregate_interventions}
\title{Aggregate Interventions}
\usage{
aggregate_interventions(conn, verbose = TRUE, render_sql = TRUE)
}
\description{
Aggregate all the Interventions from the Interventions and Intervention Other Names Table to inspect whether they are valid synonyms.
}
