% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_aact.R
\name{connect_aact}
\alias{connect_aact}
\title{Connect to AACT database}
\usage{
connect_aact(user, password = NULL)
}
\arguments{
\item{user}{Character. AACT username.}

\item{password}{Character. Default is NULL. If not provided, password will be searched for in \code{keyring} under the \code{aact} service with the provided \code{username}. If no password found, user will be interactively asked and input will be stored in keyring.}
}
\value{
\<PostgreSQLConnection\> object from \code{dbConnect}
}
\description{
Based on https://aact.ctti-clinicaltrials.org/r.
Use account credentials created via https://aact.ctti-clinicaltrials.org/users/sign_up
}
