% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_aact.R
\name{download_aact}
\alias{download_aact}
\title{Get AACT tables for given NCT ids}
\usage{
download_aact(
  ids,
  dir = here::here("raw"),
  user,
  password = NULL,
  tables = NULL,
  overwrite = FALSE,
  query = "AACT"
)
}
\arguments{
\item{ids}{Character. Vector of NCT ids to query.}

\item{dir}{Character. Directory in which to save raw CSVs. Defaults to working directory with \code{here::here("raw")}.}

\item{user}{Character. AACT username.}

\item{password}{Character. Default is NULL. If not provided, password will be searched for in \code{keyring} under the \code{aact} service with the provided \code{username}. If no password found, user will be interactively asked and input will be stored in keyring.}

\item{tables}{Character. Vector of AACT tables to query. Default of \code{NULL} results in querying all tables ("studies", "designs", "interventions", "references", "ids", "centers", "officials", "responsible-parties", "sponsors", "facilities", "central-contacts", "facility-contacts", "result-contacts").}

\item{overwrite}{Logical. If all tables already downloaded, should they be overwritten? Defaults to \code{FALSE}. Note that if \emph{any} table is missing, \emph{all} tables will be queried.}

\item{query}{Character. Query \code{INFO} for \code{loggit}. Defaults to "AACT".}
}
\description{
Calls \code{connect_aact} to connect to AACT database and \code{query_aact} to query and download individual AACT tables. Dataframes saved to specified \code{dir} as .csv. If \code{tables} not specified, all tables (as prespecified in function and messaged to user) will be queried. Logs query date with \code{loggit}.
}
\details{
AACT documentation available in the \href{https://aact.ctti-clinicaltrials.org/schema}{database schema} and \href{https://aact.ctti-clinicaltrials.org/data_dictionary}{data dictionary}.
}
\examples{
\dontrun{
download_aact(ids = c("NCT00868166", "NCT00869726"),
              dir = here::here("data", "TEST", "raw"),
              user = "respmetrics")
}
}
