% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org_trials.R
\name{get_org_trials}
\alias{get_org_trials}
\title{Get trials "led" by specified organization}
\usage{
get_org_trials(
  org,
  user,
  password = NULL,
  overwrite = FALSE,
  query = glue::glue("Organization: {org}"),
  ignore_case = TRUE
)
}
\arguments{
\item{org}{Character. Regular expression for organization.}

\item{user}{Character. AACT username.}

\item{password}{Character. Default is NULL. If not provided, password will be searched for in \code{keyring} under the \code{aact} service with the provided \code{username}. If no password found, user will be interactively asked and input will be stored in keyring.}

\item{overwrite}{Logical. If \code{org} already queried, should it be requeried? Defaults to \code{FALSE}.}

\item{query}{Character. Query \code{INFO} for \code{loggit}. Defaults to "Organization: {org}".}

\item{ignore_case}{Boolean. Whether to ignore case for \code{org}. Defaults to TRUE.}
}
\value{
Dataframe with trials with organization as lead and "leaderhip" columns
}
\description{
Trials "led" by organization if organization appears in: overall officials, sponsors (lead only), responsible parties ("principal_investigator", "study_director", "unspecified_official", "sub_investigator", "study_chair")
Based on IntoValue (https://github.com/quest-bih/IntoValue2/blob/master/code/1_sample_generation/Create_CTgov_sample.R#L87)
}
