\name{get_schema}
\alias{get_schema}
\title{Retrieve AACT Database Schema}
\description{
  Fetches and displays the schema of the AACT database.
}
\usage{
get_schema()
}
\details{
  Useful for understanding the structure of the AACT database and planning queries.
}
\value{
  Returns a representation of the AACT database schema.
}
\examples{
\dontrun{
  library(aactr)
  get_schema()
}
}
