\name{info}
\alias{info}
\title{Quick Info Reference}
\description{
  Provides basic information about the `aactR` package and how to use it.
}
\usage{
info()
}
\details{
  Displays helpful information and usage tips for the `aactR` package.
}
\value{
  Prints helpful information to the console.
}
\examples{
\dontrun{
  library(aactr)
  info()
}
}
