% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aact.R
\name{process_aact}
\alias{process_aact}
\title{Process AACT tables in given directory, process into rds}
\usage{
process_aact(
  dir_in = here::here("raw"),
  dir_out = here::here("processed"),
  overwrite = FALSE
)
}
\arguments{
\item{dir_in}{Character. Directory from which to get raw .csv files. Defaults to working directory with \code{here::here("raw")}.}

\item{dir_out}{Character. Directory in which to save processed .rds files. Defaults to working directory with \code{here::here("processed")}.}

\item{overwrite}{Logical. If all valid output tables already processed, should they be overwritten? Defaults to \code{FALSE}. Note that if \emph{any} table is not processed, \emph{all} tables will be processed. Since this checks that \emph{all} valid output tables are processed, if \emph{all} valid input tables aren't included in \code{dir_in}, \emph{all} available raw tables will be re-processed each time.}
}
\description{
Processes and cleans AACT tables from \code{dir_raw}. Dataframes saved to specified \code{dir_out} as .rds. All tables (as prespecified in function and matching \code{download_aact}) will be processed unless \emph{all} already processed and \code{overwrite} is \code{FALSE}.
}
\details{
AACT documentation available in the \href{https://aact.ctti-clinicaltrials.org/schema}{database schema} and \href{https://aact.ctti-clinicaltrials.org/data_dictionary}{data dictionary}.
}
\examples{
\dontrun{
process_aact(dir_in = here::here("data", "TEST", "raw"),
             dir_out = here::here("data", "TEST", "processed"))
}
}
