% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_aact.R
\name{query_aact}
\alias{query_aact}
\title{Query AACT table}
\usage{
query_aact(table, ids, con, filepath = NULL, ...)
}
\arguments{
\item{table}{Character. AACT table.}

\item{ids}{Character. Vector of NCT ids to query.}

\item{con}{Connection to AACT database as returned by \code{dbConnect}.}

\item{filepath}{Character. Filepath including directory portion and ending in ".csv". If no filepath provided, output will be only returned and not written to file.}

\item{...}{One or more unquoted expressions separated by commas. Column names from AACT table. \code{nct_id} is always returned and should not be included in this.}
}
\value{
Dataframe with specified columns from \code{table}
}
\description{
AACT documentation available in the \href{https://aact.ctti-clinicaltrials.org/schema}{database schema} and \href{https://aact.ctti-clinicaltrials.org/data_dictionary}{data dictionary}
}
