% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_org_in_tbl.R
\name{query_org_in_tbl}
\alias{query_org_in_tbl}
\title{Get trials which have organization in specified columns/table}
\usage{
query_org_in_tbl(org, table, columns, con, ignore_case = TRUE)
}
\arguments{
\item{org}{Character. Regular expression for organization.}

\item{table}{Character. AACT table.}

\item{columns}{Character. Vector of column names from AACT table. \code{nct_id} is always returned and should not be included in this. Organization will be searched in these columns}

\item{con}{Connection to AACT database as returned by \code{dbConnect}.}

\item{ignore_case}{Boolean. Whether to ignore case for \code{org}. Defaults to TRUE.}
}
\value{
Dataframe with organization in specified columns from \code{table}
}
\description{
AACT documentation available in the \href{https://aact.ctti-clinicaltrials.org/schema}{database schema} and \href{https://aact.ctti-clinicaltrials.org/data_dictionary}{data dictionary}
}
