% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_interventions}
\alias{search_interventions}
\title{Search Interventions}
\usage{
search_interventions(
  search_term,
  ignore.case = TRUE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE,
  conn,
  verbose = TRUE,
  render_sql = TRUE
)
}
\description{
Search an aggregate of all the interventions in the AACT database for a specific term. To retrieve the full aggregate as a dataframe, see \code{\link{aggregate_interventions}}.
}
