% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{setup_aact}
\alias{setup_aact}
\title{Setup AACT}
\usage{
setup_aact(conn, conn_fun, verbose = TRUE, render_sql = TRUE)
}
\description{
Setup AACT for the first time.
}
\seealso{
\code{\link[cli]{cat_line}}
\code{\link[rlang]{parse_expr}}
\code{\link[pg13]{dc}},\code{\link[pg13]{brake_closed_conn}},\code{\link[pg13]{lsSchema}},\code{\link[pg13]{table_exists}},\code{\link[pg13]{send}},\code{\link[pg13]{appendTable}}
\code{\link[httr]{with_config}},\code{\link[httr]{c("add_headers", "authenticate", "config", "config", "set_cookies", "timeout", "use_proxy", "user_agent", "verbose")}},\code{\link[httr]{GET}},\code{\link[httr]{content}}
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
\code{\link[secretary]{c("typewrite", "typewrite")}},\code{\link[secretary]{character(0)}}
\code{\link[curl]{handle}},\code{\link[curl]{curl_download}}
\code{\link[tibble]{tibble}}
}
