% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{update_aact}
\alias{update_aact}
\title{Setup and Maintain ClinicalTrial.gov \code{aact} Postgres Database}
\usage{
update_aact(conn, conn_fun, verbose = TRUE, render_sql = TRUE)
}
\description{
The \code{aact} Postgres database instance is available for download at \url{https://aact.ctti-clinicaltrials.org/}. Technical documentation on the AACT Database instantiation can be found at \url{https://aact.ctti-clinicaltrials.org/snapshots}. Documentation on database structure and naming conventions can be found at \url{https://aact.ctti-clinicaltrials.org/schema}.
}
\seealso{
\code{\link[cli]{cat_line}}
\code{\link[rlang]{parse_expr}}
\code{\link[pg13]{dc}},\code{\link[pg13]{brake_closed_conn}},\code{\link[pg13]{lsSchema}},\code{\link[pg13]{table_exists}},\code{\link[pg13]{send}},\code{\link[pg13]{appendTable}}
\code{\link[httr]{with_config}},\code{\link[httr]{c("add_headers", "authenticate", "config", "config", "set_cookies", "timeout", "use_proxy", "user_agent", "verbose")}},\code{\link[httr]{GET}},\code{\link[httr]{content}}
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
\code{\link[secretary]{c("typewrite", "typewrite")}},\code{\link[secretary]{character(0)}}
\code{\link[curl]{handle}},\code{\link[curl]{curl_download}}
\code{\link[tibble]{tibble}}
}
