% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rds_csv.R
\name{write_rds_csv}
\alias{write_rds_csv}
\title{Write a data frame to RDS or CSV}
\usage{
write_rds_csv(x, file)
}
\arguments{
\item{x}{A data frame or tibble to write to disk.}

\item{file}{Filepath to write to, with extension.}
}
\description{
Write a data frame to RDS or CSV
}
\examples{
\dontrun{
# Write test.rds within working directory
write_rds_csv(mtcars, fs::path("test", ext = "rds"))

# Write test.csv within test directory (which must already exist) in working directory
write_rds_csv(mtcars, fs::path("test", "test", ext = "csv"))
}
}
