% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aact_connect.R
\name{aact_connect}
\alias{aact_connect}
\title{Connect to AACT database.}
\usage{
aact_connect(user, password)
}
\arguments{
\item{user}{A character vector containing username.}

\item{password}{A character vector containing password.}
}
\description{
According to \href{https://aact.ctti-clinicaltrials.org/r}{this AACT
documentation}, this function allows you to access the PostgreSQL AACT
database in the cloud using R without needing to install PostgreSQL on your
local machine.
}
\examples{
\dontrun{
aact_connect("myid", "mypassword")
}

}
\seealso{
\url{https://aact.ctti-clinicaltrials.org/points_to_consider}
}
\author{
Jason Baik <joonwoob@andrew.cmu.edu>
}
