% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aact_connect.R
\name{aact_connect}
\alias{aact_connect}
\title{Connect to AACT database.}
\usage{
aact_connect(user, password)
}
\arguments{
\item{user}{A character vector containing username: \code{config$user}}

\item{password}{A character vector containing password: \code{config$password}}
}
\description{
According to \href{https://aact.ctti-clinicaltrials.org/r}{this AACT
documentation}, this function allows you to access the PostgreSQL AACT
database in the cloud using R without needing to install PostgreSQL on your
local machine.
}
\details{
First, create a config yaml file to securely save credentials: \code{yaml::write_yaml(data.frame(user = "USERNAME", password = "PASSWORD"), "config")}
Then, load config file and save to an object: \code{config <- yaml::yaml.load_file("config")}
}
\examples{
\dontrun{
aact_connect(config$user, config$password)
}

}
\seealso{
\url{https://aact.ctti-clinicaltrials.org/points_to_consider}
}
\author{
Jason Baik <joonwoob@andrew.cmu.edu>
}
