% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_aact.R
\name{connect_aact}
\alias{connect_aact}
\title{Connect to PostgreSQL AACT database}
\usage{
connect_aact(user, password)
}
\arguments{
\item{user}{a character vector containing username}

\item{password}{a character vector containing password}
}
\description{
According to \href{https://aact.ctti-clinicaltrials.org/r}{AACT Documentation on Using R}, this function allows you to access the PostgreSQL AACT database in the cloud using R without needing to install PostgreSQL on your local machine.
}
\examples{
\dontrun{
connect_act("myid", "mypassword")
}

}
\seealso{
\url{https://aact.ctti-clinicaltrials.org/points_to_consider}
}
\author{
Jason Baik <joonwoob@andrew.cmu.edu>
}
